requirements.txt
pyupbit==0.2.33
pandas==2.1.4
numpy==1.26.2
ta==0.11.0
schedule==1.2.1
python-telegram-bot==20.7
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
scikit-learn==1.3.2
